Trail implemented, particle system revamped
The trail now only shows when drifting. It is using a custom shader graph to improve the overall look and add a fade to the colour.

The particle system has been tidied up to now have less particles in a concentrated area so it looks contained to the battery. It also fully disappears when not at 100% making the full charged shot look like it's depleating it.